feat(sense): restructure avatar menu (shadcn) + accent menu highlight
Avatar menu extracted to SiteHeader/UserAvatarMenu.tsx and rebuilt to shadcn
structure — every desktop row is a keyboard-reachable menu item:

- Profile/org rows, actions, and legal links are DropdownMenuItems (desktop) /
  buttons (mobile sheet). Feature Requests uses the new DropdownMenuLinkItem;
  Log out stays a menu item (client action).
- Legal links open sense Dialogs via lifted state (rendered outside the menu),
  scrollable-body pattern, initialFocus pinned to the top for a11y.
- Mobile sheet now scrolls (flex column + overflow body).

@op/sense component changes it needed:
- New DropdownMenuLinkItem (base-ui Menu.LinkItem) so in-menu links join roving
  focus instead of dragging the open scroll to the bottom.
- DropdownMenuSeparator gets shrink-0 (flex menus were collapsing it to 0px).
- Menu/select item highlight -> accent (teal-50) across dropdown-menu,
  context-menu, menubar, select, command (reverses the P1 muted choice;
  NavigationMenu left as-is).
- ProfileItem: min-w-0 + truncate so long names ellipsize.
