+++ apps/app/src/components/PrivacyPolicyModal/index.tsx
@@ -16,7 +16,7 @@ export const PrivacyPolicyModal = () => {
       <Button
         unstyled
         onPress={() => setIsOpen(true)}
-        className="text-primary-teal hover:underline"
+        className="text-sm font-strong text-primary-teal hover:underline"
       >
         {t('Privacy Policy')}
       </Button>
