+++ apps/app/src/components/ToSModal/index.tsx
@@ -11,7 +11,10 @@ export const ToSModal = () => {
 
   return (
     <DialogTrigger>
-      <Button unstyled className="text-primary-teal hover:underline">
+      <Button
+        unstyled
+        className="text-sm font-strong text-primary-teal hover:underline"
+      >
         {t('Terms of Service')}
       </Button>
